added purge.js, fixed say.js

+++ cmds/say.js
@@ -1,6 +1,8 @@
 module.exports.run = async (client, message, args) => {
 	const sayMessage = args.join(" ");
 
+	if(!sayMessage) return message.reply("You didn't type anything")
+
 	message.delete().catch(error=>{}); 
 	message.channel.send(sayMessage);
 }
